+++ tests/ConfigContraband.Tests/Infrastructure/Verifier.cs
@@ -225,6 +225,45 @@ internal static class Verifier
         await test.RunAsync();
     }
 
+    public static Task VerifyCodeFixAsync(
+        (string filename, string content)[] sources,
+        (string filename, string content)[] fixedSources,
+        (string filename, string content) additionalFile,
+        params DiagnosticResult[] expected)
+    {
+        return VerifyCodeFixAsync(sources, fixedSources, Microsoft.CodeAnalysis.OutputKind.DynamicallyLinkedLibrary, additionalFile, expected);
+    }
+
+    public static async Task VerifyCodeFixAsync(
+        (string filename, string content)[] sources,
+        (string filename, string content)[] fixedSources,
+        Microsoft.CodeAnalysis.OutputKind outputKind,
+        (string filename, string content) additionalFile,
+        params DiagnosticResult[] expected)
+    {
+        var test = new CSharpCodeFixTest<ConfigContrabandAnalyzer, ConfigContrabandCodeFixProvider, DefaultVerifier>
+        {
+            ReferenceAssemblies = OptionsReferences
+        };
+
+        foreach (var source in sources)
+        {
+            test.TestState.Sources.Add(source);
+        }
+
+        foreach (var fixedSource in fixedSources)
+        {
+            test.FixedState.Sources.Add(fixedSource);
+        }
+
+        test.TestState.OutputKind = outputKind;
+        test.FixedState.OutputKind = outputKind;
+        test.TestState.AdditionalFiles.Add(additionalFile);
+
+        test.ExpectedDiagnostics.AddRange(expected);
+        await test.RunAsync();
+    }
+
     public static async Task VerifyFixAllAsync(
         string source,
         string fixedSource,